requirements.txt
streamlit==1.13.0
pandas==1.4.2
numpy==1.21.6
matplotlib==3.5.1
statsmodels==0.13.2
scikit-learn==1.0.2
gspread==5.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6